+++ DatadogRUM/Sources/Instrumentation/AppHangs/AppHang.swift
@@ -19,6 +19,9 @@ internal struct AppHang: Codable {
         /// Indicates that backtrace generation is unavailable.
         /// It is the case when `BacktraceReportingFeature` is not available in core (when Crash Reporting feature was not enabled).
         case notAvailable
+        /// Indicates that backtrace generation was turned off in `CrashReporting.Configuration`.
+        /// Unlike `notAvailable`, Crash Reporting **is** enabled - it was asked not to provide App Hang backtraces.
+        case disabled
     }
 
     /// The date of hang start.