+++ DatadogRUM/Tests/Instrumentation/AppHangs/AppHangsMonitorTests.swift
@@ -87,6 +87,27 @@ class AppHangsMonitorTests: XCTestCase {
         XCTAssertEqual(command.isStackTraceTruncated, hang.backtraceResult.wasTruncated)
     }
 
+    func testWhenAppHangEndsWithBacktraceGenerationDisabled_itSendsAppHangCommandWithNoStackTrace() throws {
+        // Given
+        let subscriber = RUMCommandSubscriberMock()
+        monitor.nonFatalHangsHandler.publish(to: subscriber)
+        monitor.start()
+        defer { monitor.stop() }
+
+        // When
+        let hang: AppHang = .mockWith(backtraceResult: .disabled)
+        watchdogThread.delegate?.hangEnded(hang, duration: .mockRandom(min: 1, max: 4))
+
+        // Then
+        let command = try XCTUnwrap(subscriber.lastReceivedCommand as? RUMAddCurrentViewAppHangCommand)
+        XCTAssertEqual(command.message, AppHangsMonitor.Constants.appHangErrorMessage)
+        XCTAssertEqual(command.type, AppHangsMonitor.Constants.appHangErrorType)
+        XCTAssertEqual(command.stack, AppHangsMonitor.Constants.appHangStackDisabledErrorMessage)
+        XCTAssertNil(command.threads)
+        XCTAssertNil(command.binaryImages)
+        XCTAssertNil(command.isStackTraceTruncated)
+    }
+
     // MARK: - Fatal App Hangs Monitoring
 
     func testGivenFatalErrorViewContextAvailable_whenAppHangStarts_itSavesPendingAppHangToDataStore() throws {